+++ CHANGELOG.md
@@ -56,6 +56,11 @@ to ParallelExecutor, mutex state lock.
   must be removed or replaced with `using namespace SimTK::Xml` or
   `using SimTK::Xml::Document` depending on the intent. 
   [PR #460](https://github.com/simbody/simbody/pull/460)
+* Improved `NiceTypeName<T>::namestr()` to produce a canonicalized name that is
+  the same on all platforms (with a few exceptions). Added `xmlstr()` method to
+  make an XML-friendly modification of `namestr()` that replaces angle brackets
+  with curly braces. Added a new regression test to verify that the names come
+  out right. [PR #461](https://github.com/simbody/simbody/pull/461)
 * (There are more that haven't been added yet)